+++ src/DotNetOutdated/Constants.cs
@@ -1,0 +1,10 @@
+﻿namespace DotNetOutdated
+{
+    public static class Constants
+    {
+        public static class Reporting
+        {
+            public const string UnknownValue = "-";
+        }
+    }
+}