+++ src/DotNetOutdated/ReportingExtensions.cs
@@ -9,8 +9,8 @@ namespace DotNetOutdated
         {
             List<int> columnWidths = new List<int>();
             columnWidths.Add(packages.Select(p => p.Name).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length);
-            columnWidths.Add(packages.Select(p => p.ReferencedVersion.ToString()).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length);
-            columnWidths.Add(packages.Select(p => p.LatestVersion.ToString()).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length);
+            columnWidths.Add(packages.Select(p => p.ReferencedVersion?.ToString() ?? Constants.Reporting.UnknownValue).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length);
+            columnWidths.Add(packages.Select(p => p.LatestVersion?.ToString() ?? Constants.Reporting.UnknownValue).Aggregate("", (max, cur) => max.Length > cur.Length ? max : cur).Length);
 
             return columnWidths.ToArray();
         }