+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [v0.2.0]
 - Display column headers ([#1](https://github.com/jerriep/dotnet-outdated/issues/1))
 - Allow specifying whether to include pre-release versions ([#4](https://github.com/jerriep/dotnet-outdated/issues/4))
 - Fix bug when latest version cannot be found ([#10](https://github.com/jerriep/dotnet-outdated/issues/10))
@@ -9,4 +9,5 @@
 Initial release
  - A .NET Core global tool to display outdated NuGet packages in a project
 
+[v0.2.0]: https://github.com/jerriep/dotnet-outdated/tree/v0.2.0
 [v0.1.0]: https://github.com/jerriep/dotnet-outdated/tree/v0.1.0